+++ CHANGELOG.md
@@ -11,6 +11,7 @@ adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 - A `disconnect` signal to let Infinity know that the `MediaConnection` is going away
 - `ConferenceStep.conference(String)` to create a new `ConferenceStep` using a different alias
+- `VideoTrackEmbeddedRenderer` as an alternative to `VideoTrackRenderer`
 
 ### Changed